vertex_bq_vector_demo/requirements.txt
google-cloud-aiplatform>=1.70.0
google-cloud-bigquery>=3.25.0
pandas>=2.0.0
vertexai>=1.71.1
